% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbarima.R
\name{rbarima}
\alias{rbarima}
\title{Rbarima}
\usage{
rbarima(df = .Last.value, Measure = "", Time = "", Pred.time = "",
  launch.browser = False)
}
\arguments{
\item{df}{A dataframe input}

\item{Measure}{Specify value measures such as Revenue and Shipments}

\item{Time}{Specify time variables such as Year, Quarter, and Month}

\item{Pred.time}{Specify a number of years or quarters to predict}
}
\description{
Rbarima
}
